+++ src/azure-cli/azure/cli/command_modules/network/aaz/latest/network/public_ip/ddos_protection_statu/__cmd_group.py
@@ -12,10 +12,10 @@ from azure.cli.core.aaz import *
 
 
 @register_command_group(
-    "network public-ip ddos-protection-statu",
+    "network public-ip ddos-protection-status",
 )
 class __CMDGroup(AAZCommandGroup):
-    """Manage public IP ddos protection statu.
+    """Manage public IP ddos protection status.
     """
     pass